Autenticação parte 1

+++ routes/web.php
@@ -1,36 +1,23 @@
 <?php
 
-use App\Http\Controllers\{
-    UserController,
-    ViaCepController,
-    PostController
-};
 use Illuminate\Support\Facades\Route;
+use App\Http\Controllers\UserController;
 
-Route::redirect('/', '/users');
+require __DIR__.'/auth.php';
 
-// Route::get('/', function () {
-//     return view('welcome');
-// });
+Route::get('/', function() {
+    return view('home');
+});
 
-Route::get('/posts', [PostController::class, 'index'])->name('posts.index');
+Route::middleware(['auth'])->group(function () {
+    Route::delete('/users/{id}', [UserController::class, 'destroy'])->name('users.destroy')->middleware('auth');
+    Route::put('users/{id}', [UserController::class, 'update'])->name('users.update')->middleware('auth');
+    Route::get('/users/{id}/edit', [UserController::class, 'edit'])->name('users.edit')->middleware('auth');
+    Route::get('/users', [UserController::class, 'index'])->name('users.index')->middleware('auth');
+    Route::get('/users/create', [UserController::class, 'create'])->name('users.create')->middleware('auth');
+    Route::post('/user', [UserController::class, 'store'])->name('users.store');
+    Route::get('/users/{id}', [UserController::class, 'show'])->name('users.show');
 
-
-Route::get('/users/{id}/posts', [PostController::class, 'show'])->name('posts.show');
-
-// Route::get('/users/{nome}', function ($nome) {
-//     echo $nome;
-// });
-
-Route::delete('/users/{id}', [UserController::class, 'destroy'])->name('users.destroy');
-Route::put('users/{id}', [UserController::class, 'update'])->name('users.update');
-Route::get('/users/{id}/edit', [UserController::class, 'edit'])->name('users.edit');
-Route::get('/users/create', [UserController::class, 'create'])->name('users.create');
-Route::post('/user', [UserController::class, 'store'])->name('users.store');
-Route::get('/users/{id}', [UserController::class, 'show'])->name('users.show');
-Route::get('/users', [UserController::class, 'index'])->name('users.index');
-
-// VIA CEP WEB SERVICE
-Route::get('/viacep', [ViaCepController::class, 'index'])->name('viacep.index');
-Route::post('/viacep', [ViaCepController::class, 'index'])->name('viacep.index.post');
-Route::get('/viacep/{cep}', [ViaCepController::class, 'show'])->name('viacep.show');
+    Route::get('/posts', [PostController::class, 'index'])->name('posts.index');
+    Route::get('/users/{id}/posts', [PostController::class, 'show'])->name('posts.show');
+});
